+++ pybamm/expression_tree/operations/latexify.py
@@ -2,7 +2,6 @@
 # Latexify class
 #
 import copy
-import operator
 import re
 import warnings
 
@@ -55,58 +54,6 @@ class Latexify:
         self.filename = filename
         self.newline = newline
 
-    def _get_concat_displays(self, node):
-        """
-        Returns all the concatenation nodes by doing a depth first search through the
-        entire equation tree with ranges in front of all nodes.
-        """
-        concat_displays = []
-        dfs_nodes = [node]
-        while dfs_nodes:
-            node = dfs_nodes.pop()
-            if (
-                hasattr(node, "concat_latex")
-                and getattr(node, "print_name", None) is not None
-            ):
-                # Combine list of concatenations with list of ranges
-                concat_geo = map(
-                    operator.add,
-                    node.concat_latex,
-                    self._get_concat_geometry_displays(node),
-                )
-
-                # Add cases and split by new line
-                concat_sym = (
-                    r"\begin{cases}" + r" \\ ".join(concat_geo) + r"\end{cases}"
-                )
-                concat_eqn = sympy.Eq(
-                    sympy.Symbol(node.print_name),
-                    sympy.Symbol(concat_sym),
-                    evaluate=False,
-                )
-                concat_displays.append(concat_eqn)
-            dfs_nodes.extend(node.children)
-
-        # Remove duplicates from the list whilst preserving order
-        return list(dict.fromkeys(concat_displays))
-
-    def _get_concat_geometry_displays(self, var):
-        """Returns a list of min/max ranges of all concatenation nodes in latex."""
-        geo = []
-
-        # Loop through all subdomains for concatenations
-        for domain in var.domain:
-            for var_name, rng in self.model.default_geometry[domain].items():
-                if "min" in rng and "max" in rng:
-                    rng_min = get_rng_min_max_name(rng, "min")
-                    rng_max = get_rng_min_max_name(rng, "max")
-
-                    name = sympy.latex(var_name)
-                    geo_latex = f"& {rng_min} < {name} < {rng_max}"
-                    geo.append(geo_latex)
-
-        return geo
-
     def _get_geometry_displays(self, var):
         """
         Returns min range from the first domain and max range from the last domain of
@@ -136,7 +83,7 @@ class Latexify:
 
         return geo
 
-    def _get_bcs_displays(self, lhs_dr, var):
+    def _get_bcs_displays(self, var):
         """
         Returns a list of boundary condition equations with ranges in front of
         the equations.
@@ -146,27 +93,25 @@ class Latexify:
 
         if bcs:
             # Take range minimum from the first domain
-            for var_name, rng in self.model.default_geometry[var.domain[0]].items():
-                # Trim name (r_n --> r)
-                name = re.findall(r"(.)_*.*", str(var_name))[0]
-                rng_min = get_rng_min_max_name(rng, "min")
+            var_name = list(self.model.default_geometry[var.domain[0]].keys())[0]
+            rng_left = list(self.model.default_geometry[var.domain[0]].values())[0]
+            rng_right = list(self.model.default_geometry[var.domain[-1]].values())[0]
 
-                bcs_left = sympy.latex(bcs["left"][0].to_equation())
-                bcs_left_latex = bcs_left + f"\quad {name} = {rng_min}"
-                bcs_eqn = sympy.Eq(lhs_dr, sympy.Symbol(bcs_left_latex), evaluate=False)
-                bcs_eqn_list.append(bcs_eqn)
+            # Trim name (r_n --> r)
+            var_name = re.findall(r"(.)_*.*", str(var_name))[0]
 
-            # Take range maximum from the last domain
-            for var_name, rng in self.model.default_geometry[var.domain[-1]].items():
-                # Trim name (r_n --> r)
-                name = re.findall(r"(.)_*.*", str(var_name))[0]
-                rng_max = get_rng_min_max_name(rng, "max")
+            rng_min = get_rng_min_max_name(rng_left, "min")
+            rng_max = get_rng_min_max_name(rng_right, "max")
 
-                bcs_right = sympy.latex(bcs["right"][0].to_equation())
-                bcs_right_latex = bcs_right + f"\quad {name} = {rng_max}"
-                bcs_eqn = sympy.Eq(
-                    lhs_dr, sympy.Symbol(bcs_right_latex), evaluate=False
-                )
+            for side, rng in [("left", rng_min), ("right", rng_max)]:
+                bc_value, bc_type = bcs[side]
+                bcs_side = sympy.latex(bc_value.to_equation())
+                bcs_side_latex = bcs_side + f"\\quad  \\text{{at }} {var_name} = {rng}"
+                if bc_type == "Dirichlet":
+                    lhs = sympy.Symbol(var.print_name)
+                else:
+                    lhs = sympy.Symbol(r"\nabla " + var.print_name)
+                bcs_eqn = sympy.Eq(lhs, sympy.Symbol(bcs_side_latex), evaluate=False)
                 bcs_eqn_list.append(bcs_eqn)
 
         return bcs_eqn_list
@@ -214,7 +159,14 @@ class Latexify:
 
         return param_list, var_list
 
-    def latexify(self):
+    def latexify(self, output_variables=None):
+        # Voltage is the default output variable if it exists
+        if output_variables is None:
+            if "Voltage [V]" in self.model.variables:
+                output_variables = ["Voltage [V]"]
+            else:
+                output_variables = []
+
         eqn_list = []
         param_list = []
         var_list = []
@@ -234,22 +186,24 @@ class Latexify:
                 eqn_list.append(sympy.Symbol(r"\\ \textbf{" + str(var) + "}"))
 
                 # Set lhs derivative
-                lhs = sympy.Derivative(var_symbol, "t")
-                lhs_dr = sympy.Derivative(var_symbol, "r")
+                ddt = sympy.Derivative(var_symbol, "t")
 
                 # Override lhs for algebraic
-                if eqn_type == "algebraic":
+                if eqn_type == "rhs":
+                    lhs = ddt
+                else:
                     lhs = 0
 
                 # Override derivative to partial derivative
-                if len(var.domain) != 0 and var.domain != "current collector":
-                    lhs_dr.force_partial = True
-
-                    if not eqn_type == "algebraic":
-                        lhs.force_partial = True
+                if (
+                    len(var.domain) != 0
+                    and var.domain != "current collector"
+                    and eqn_type == "rhs"
+                ):
+                    lhs.force_partial = True
 
                 # Boundary conditions equations
-                bcs = self._get_bcs_displays(lhs_dr, var)
+                bcs = self._get_bcs_displays(var)
 
                 # Add ranges from geometry in rhs
                 geo = self._get_geometry_displays(var)
@@ -263,13 +217,12 @@ class Latexify:
                 if not eqn_type == "algebraic":
                     init = self.model.initial_conditions.get(var, None)
                     init_eqn = sympy.Eq(var_symbol, init.to_equation(), evaluate=False)
-                    init_eqn = sympy.Symbol(sympy.latex(init_eqn) + r"\quad at\; t=0")
+                    init_eqn = sympy.Symbol(
+                        sympy.latex(init_eqn) + r"\quad \text{at}\; t=0"
+                    )
 
                 # Make equation from lhs and rhs
                 lhs_rhs = sympy.Eq(lhs, rhs, evaluate=False)
-
-                # Get all concatenation nodes
-                concat_displays = self._get_concat_displays(eqn)
 
                 # Set SymPy's init printing to use CustomPrint from sympy_overrides.py
                 sympy.init_printing(
@@ -281,10 +234,6 @@ class Latexify:
 
                 # Add model equations to the list
                 eqn_list.append(lhs_rhs)
-
-                # Add concatenation to the list
-                if concat_displays:
-                    eqn_list.append(concat_displays)
 
                 # Add initial conditions to the list
                 if not eqn_type == "algebraic":
@@ -298,23 +247,21 @@ class Latexify:
                 param_list.extend(list1)
                 var_list.extend(list2)
 
-        # Add voltage expression to the list
-        if "Voltage [V]" in self.model.variables:
-            voltage = self.model.variables["Voltage [V]"].to_equation()
-            voltage_eqn = sympy.Eq(sympy.Symbol("V"), voltage, evaluate=False)
-            # Add voltage to the list
-            eqn_list.append(sympy.Symbol(r"\\ \textbf{Voltage [V]}"))
-            eqn_list.extend([voltage_eqn])
+        # Add output variables to the list
+        for var_name in output_variables:
+            var = self.model.variables[var_name].to_equation()
+            var_eqn = sympy.Eq(sympy.Symbol("V"), var, evaluate=False)
+            # Add var to the list
+            eqn_list.append(sympy.Symbol(r"\\ \textbf{" + var_name + "}"))
+            eqn_list.extend([var_eqn])
 
         # Remove duplicates from the list whilst preserving order
         param_list = list(dict.fromkeys(param_list))
         var_list = list(dict.fromkeys(var_list))
         # Add Parameters and Variables to the list
         eqn_list.append(sympy.Symbol(r"\\ \textbf{Parameters and Variables}"))
-        # Add parameters to the list
-        eqn_list.extend(param_list)
-        # Add names to the list
         eqn_list.extend(var_list)
+        eqn_list.extend(param_list)
 
         # Split list with new lines
         eqn_new_line = sympy.Symbol(r"\\\\".join(map(custom_print_func, eqn_list)))