+++ pybamm/parameters/base_parameters.py
@@ -100,14 +100,7 @@ class BaseParameters:
 class NullParameters:
     def __getattribute__(self, name):
         "Returns 0 for some parameters that aren't found by __getattribute__"
-        if name in [
-            "epsilon_s",
-            "Q_init",
-            "n_Li_init",
-            "Q_Li_init",
-            "R_typ",
-            "j_scale",
-        ]:
+        if name in ["epsilon_s", "Q_init", "n_Li_init", "Q_Li_init", "R_typ"]:
             return pybamm.Scalar(0)
         else:
             return super().__getattribute__(name)