+++ pybamm/simulation.py
@@ -392,14 +392,10 @@ class Simulation:
         if self.model_with_set_params:
             return
 
-        if self._parameter_values._dict_items == {}:
-            # Don't process if parameter values is empty
-            self._model_with_set_params = self._unprocessed_model
-        else:
-            self._model_with_set_params = self._parameter_values.process_model(
-                self._unprocessed_model, inplace=False
-            )
-            self._parameter_values.process_geometry(self.geometry)
+        self._model_with_set_params = self._parameter_values.process_model(
+            self._unprocessed_model, inplace=False
+        )
+        self._parameter_values.process_geometry(self.geometry)
         self.model = self._model_with_set_params
 
     def set_initial_soc(self, initial_soc):