+++ tests/unit/test_util.py
@@ -108,7 +108,7 @@ class TestSearch(unittest.TestCase):
         model = pybamm.BaseModel()
         model.variables = {"Electrolyte concentration": 1, "Electrode potential": 0}
 
-        param = pybamm.ParameterValues({"a": 10, "b": 2})
+        param = pybamm.ParameterValues({"test": 10, "b": 2})
 
         # Test variables search (default returns key)
         with patch("sys.stdout", new=StringIO()) as fake_out:
@@ -127,8 +127,8 @@ class TestSearch(unittest.TestCase):
 
         # Test param search (default returns key, value)
         with patch("sys.stdout", new=StringIO()) as fake_out:
-            param.search("a")
-            self.assertEqual(fake_out.getvalue(), "a\t10\n")
+            param.search("test")
+            self.assertEqual(fake_out.getvalue(), "test\t10\n")
 
 
 if __name__ == "__main__":